# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AFRIGAZ",
        "description": "Documentation API AFRIGAZ",
        "contact": {
            "email": "winnersthec001.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/accountStoreData": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Créer",
                "operationId": "457e59b7bc2496b186f8fddb20bd9359",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "designation",
                                    "branche_id",
                                    "nature"
                                ],
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "branche_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nature": {
                                        "type": "string",
                                        "example": "Nature du compte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Données créées avec succès"
                    },
                    "422": {
                        "description": "Validation des données échouée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/v1/accountUpdate/{id}": {
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Modifier",
                "operationId": "993986a004d952c61a807ac1783f515e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "designation": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "branche_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nature": {
                                        "type": "string",
                                        "example": "Nature du compte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "mise à jour"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
        "/api/v1/accountDelete/{id}": {
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Supprimer",
                "operationId": "17fae7095e2b1b6f6c0cc5b7d5e0420f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supprimée"
                    },
                    "404": {
                        "description": "Non trouvée"
                    }
                }
            }
        },
...
    },
...
}